Amend Task definition
Tasks have new attributes which specify whether they are
repeating and, if so, with what period, whether it is useful to
partially complete them and the maximum number of time chunks into
which they can be divided. Tasks also have a UID associated with
them.

Additionally, the time property of the class has been changed back
to a change_time method since the level of handling required on the
possible input parameters was beyond what is acceptable for an
attribute.