Add JSON encoder and decoder classes to TimeChunk classes in time.time.py
TimeChunk and AllocatedTimeChunk now have JSON encoder and decoder
classes which rely on Datetime and Timedelta implementations in
time.util.

Instead of importing the time.util module as a whole, thread.task
now imports just its classes.

Also, a function parameter name in time.util.py has been changed
from s to f to reflect that it is a float and not a string.