Debug and refactor packages, add doctests to threads/task.py
The two packages were tested. time conflicted with the builtin
time module and so is now renamed to timemap. the thread package
has been renamed as threads. Additionally, small errors in
timemap/util.py and timemap/time.py have been rectified.

A subset of the tests conducted are now placed in doctests of
threads/task.py.