Add RepeatableTask and inherit Event and Assignment from it
The RepeatableTask class with its TaskRepeat object is able to
handle repeating tasks. When subclassed in Event and Assignment, it
is able to handle multiple appointments and deadlines per repeat
period respectively.

The classes also have new JSON encoders and decoders for accurately
storing the state of the class. They have been created keeping
time inspection in mind.